requirements.txt
flask
python-dotenv
requests
eth-brownie